move translator mode settings to separate panel
