+++ src/components/settings/UISettings.svelte
@@ -6,7 +6,6 @@
     chatZoom,
     livetlFontSize,
     doSpeechSynth,
-    doTranslatorMode,
     showCaption,
     showTimestamp,
     speechVolume,
@@ -22,7 +21,6 @@
   import EnumOption from '../options/Radio.svelte';
   import FontDemo from '../FontDemo.svelte';
   import ImportExport from '../ImportExport.svelte';
-  import TranslatorMode from '../options/TranslatorMode.svelte';
   export let isStandalone = false;
 </script>
 
@@ -90,7 +88,3 @@
   name="Show fullscreen button"
   store={enableFullscreenButton}
 />
-<CheckOption name="Translator mode" store={doTranslatorMode} />
-{#if $doTranslatorMode}
-  <TranslatorMode />
-{/if}
